CB_UserBased/content-based-algorithm/src/requirements.txt
certifi
chardet
click
idna
isodate
joblib
nltk
pandas
numpy
python-dateutil
pytz
PyYAML
redis
regex
rejson
requests
scikit-learn
scipy
six
threadpoolctl
tqdm
urllib3
APScheduler